[Build] Fixed MSVC build

+++ Source/Test/Automation/Include/Test/Automation/Connection/PoolingListener.h
@@ -48,21 +48,21 @@ private:
     /// \param ptr data begin
     /// \param length byte length of data
     /// \param count number of messages
-    void Commit(MessageSchema schema, const void* ptr, size_t length, uint32_t count);
+    void Commit(MessageSchema schema, const void* ptr, size_t length, size_t count);
 
     /// Commit all pending tasks for a bucket
     /// \param bucket bucket to apply for
     /// \param ptr data begin
     /// \param length byte length of data
     /// \param count number of messages
-    void Commit(TaskBucket& bucket, const void* ptr, size_t length, uint32_t count);
+    void Commit(TaskBucket& bucket, const void* ptr, size_t length, size_t count);
 
     /// Commit all pending tasks for a task
     /// \param task task to apply for
     /// \param ptr data begin
     /// \param length byte length of data
     /// \param count number of messages
-    void Commit(MessageTask* task, const void* ptr, size_t length, uint32_t count);
+    void Commit(MessageTask* task, const void* ptr, size_t length, size_t count);
 
 private:
     struct SchemaComparator {
